refactor: remove `GeniusYield.Orderbot.Types` module
